Factor our AffineCheck.{h,cpp}
